Feat: add simple player movement

+++ src/entity/Entity.java
@@ -6,18 +6,18 @@ public abstract class Entity {
     public Entity() {}
 
     // representa a posição x e y, de uma forma simplificada
-    protected int m_Position[] = {0, 0};
+    protected float m_Position[] = {0, 0};
 
     public abstract void update();
     public abstract void render();
 
     public int getXPosition()
     {
-        return m_Position[0];
+        return (int)m_Position[0];
     }
 
     public int getYPosition()
     {
-        return m_Position[1];
+        return (int)m_Position[1];
     }
 }
